requirements.txt
streamlit
pandas
python-docx
Pillow
openpyxl
